service-vendor-aws-api/requirements.txt
appdirs==1.4.4
attrs==20.3.0
beautifulsoup4==4.9.3
boto3==1.16.63
botocore==1.19.63
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
distlib==0.3.1
filelock==3.0.12
identify==2.2.1
idna==2.10
importlib-metadata==3.7.3
jmespath==0.10.0
jsonschema==3.2.0
lxml==4.6.2
nodeenv==1.5.0
pre-commit==2.11.1
pyaml==20.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
soupsieve==2.1
SQLAlchemy==1.4.19
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.4.3
deepdiff==5.5.0
xmljson==0.2.1
zipp==3.4.1
PyMySQL==1.0.2
pytz==2021.1
pytest==6.2.4
pytest-cov==2.12.1
dnspython >= 2.0.0
